write the new test case to replace the old noe.
